+++ common/cards/alter-egos-ii/hermits/architectfalse-common.ts
