+++ common/cards/alter-egos-ii/hermits/fiveampearl-rare.ts
