+++ common/cards/alter-egos-ii/hermits/frenchkeralis-common.ts
