+++ common/cards/alter-egos-ii/hermits/index.ts
@@ -1,0 +1,34 @@
+import HermitCard from '../../base/hermit-card'
+import ArchitectFalseCommonHermitCard from '../../alter-egos-ii/hermits/architectfalse-common'
+import BoomerBdubsRareHermitCard from '../../alter-egos-ii/hermits/boomerbdubs-rare'
+import DwarfImpulseCommonHermitCard from '../../alter-egos-ii/hermits/dwarfimpulse-common'
+import FiveAMPearlRareHermitCard from '../../alter-egos-ii/hermits/fiveampearl-rare'
+import FrenchKeralisCommonHermitCard from '../../alter-egos-ii/hermits/frenchkeralis-common'
+import HorseHeadHypnoCommonHermitCard from '../../alter-egos-ii/hermits/horseheadhypno-rare'
+import IskallmanRareHermitCard from '../../alter-egos-ii/hermits/iskallman-rare'
+import OriginalXbCommonHermitCard from '../../alter-egos-ii/hermits/originalxb-common'
+import OverseerRareHermitCard from '../../alter-egos-ii/hermits/overseer-rare'
+import PrincessGemCommonHermitCard from '../../alter-egos-ii/hermits/princessgem-common'
+import ShadEECommonHermitCard from './shadee-common'
+import SpookyStressCommonHermitCard from '../../alter-egos-ii/hermits/spookystress-common'
+import SteampunkTangoCommonHermitCard from '../../alter-egos-ii/hermits/steampunktango-common'
+import WormManCommonHermitCard from '../../alter-egos-ii/hermits/wormman-common'
+
+const hermitCardClasses: Array<HermitCard> = [
+	new ArchitectFalseCommonHermitCard(),
+	new BoomerBdubsRareHermitCard(),
+	new DwarfImpulseCommonHermitCard(),
+	new FiveAMPearlRareHermitCard(),
+	new FrenchKeralisCommonHermitCard(),
+	new HorseHeadHypnoCommonHermitCard(),
+	new IskallmanRareHermitCard(),
+	new OriginalXbCommonHermitCard(),
+	new OverseerRareHermitCard(),
+	new PrincessGemCommonHermitCard(),
+	new ShadEECommonHermitCard(),
+	new SpookyStressCommonHermitCard(),
+	new SteampunkTangoCommonHermitCard(),
+	new WormManCommonHermitCard(),
+]
+
+export default hermitCardClasses
