+++ common/cards/alter-egos-ii/hermits/iskallman-rare.ts
