+++ common/cards/alter-egos-ii/hermits/originalxb-common.ts
