+++ common/cards/alter-egos-ii/hermits/princessgem-common.ts
