+++ common/cards/alter-egos-ii/hermits/shadee-common.ts
