+++ common/cards/alter-egos-ii/hermits/spookystress-common.ts
