+++ common/cards/alter-egos-ii/hermits/wormman-common.ts
