--- common/cards/alter-eogs-ii/hermits/index.ts
@@ -1,34 +1,0 @@
-import HermitCard from '../../base/hermit-card'
-import ArchitectFalseCommonHermitCard from '../../alter-eogs-ii/hermits/architectfalse-common'
-import BoomerBdubsRareHermitCard from '../../alter-eogs-ii/hermits/boomerbdubs-rare'
-import DwarfImpulseCommonHermitCard from '../../alter-eogs-ii/hermits/dwarfimpulse-common'
-import FiveAMPearlRareHermitCard from '../../alter-eogs-ii/hermits/fiveampearl-rare'
-import FrenchKeralisCommonHermitCard from '../../alter-eogs-ii/hermits/frenchkeralis-common'
-import HorseHeadHypnoCommonHermitCard from '../../alter-eogs-ii/hermits/horseheadhypno-rare'
-import IskallmanRareHermitCard from '../../alter-eogs-ii/hermits/iskallman-rare'
-import OriginalXbCommonHermitCard from '../../alter-eogs-ii/hermits/originalxb-common'
-import OverseerRareHermitCard from '../../alter-eogs-ii/hermits/overseer-rare'
-import PrincessGemCommonHermitCard from '../../alter-eogs-ii/hermits/princessgem-common'
-import ShadEECommonHermitCard from './shadee-common'
-import SpookyStressCommonHermitCard from '../../alter-eogs-ii/hermits/spookystress-common'
-import SteampunkTangoCommonHermitCard from '../../alter-eogs-ii/hermits/steampunktango-common'
-import WormManCommonHermitCard from '../../alter-eogs-ii/hermits/wormman-common'
-
-const hermitCardClasses: Array<HermitCard> = [
-	new ArchitectFalseCommonHermitCard(),
-	new BoomerBdubsRareHermitCard(),
-	new DwarfImpulseCommonHermitCard(),
-	new FiveAMPearlRareHermitCard(),
-	new FrenchKeralisCommonHermitCard(),
-	new HorseHeadHypnoCommonHermitCard(),
-	new IskallmanRareHermitCard(),
-	new OriginalXbCommonHermitCard(),
-	new OverseerRareHermitCard(),
-	new PrincessGemCommonHermitCard(),
-	new ShadEECommonHermitCard(),
-	new SpookyStressCommonHermitCard(),
-	new SteampunkTangoCommonHermitCard(),
-	new WormManCommonHermitCard(),
-]
-
-export default hermitCardClasses
