+++ common/cards/index.ts
@@ -13,7 +13,7 @@ import defaultItemCards from './default/items'
 import defaultSingleUseCards from './default/single-use'
 import alterEgosSingleUseCards from './alter-egos/single-use'
 import adventOfTcgSingleUseCards from './advent-of-tcg/single-use'
-import alterEgosIIHermitCards from './alter-eogs-ii/hermits/index'
+import alterEgosIIHermitCards from './alter-egos-ii/hermits/index'
 
 const effectCardClasses: Array<EffectCard> = [
 	...defaultEffectCards,
